refactoring, rename files and folders
